Cambios en el tableview modificando una celda y una clase Utils.

+++ src/dad/autoescuela/controllers/MenuProfesorController.java
@@ -1,21 +1,28 @@
 package dad.autoescuela.controllers;
 
 import java.io.File;
+import java.util.Optional;
 
-import dad.autoescuela.Main;
 import dad.autoescuela.MainProfesor;
 import dad.autoescuela.model.Pregunta;
 import dad.autoescuela.model.Usuario;
 import dad.autoescuela.resources.images.Images;
 import dad.autoescuela.services.ServiceLocator;
+import dad.autoescuela.utils.Utils;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
 import javafx.event.Event;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
+import javafx.geometry.Pos;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.Label;
 import javafx.scene.control.RadioButton;
+import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
@@ -24,9 +31,13 @@ import javafx.scene.control.TextField;
 import javafx.scene.control.Toggle;
 import javafx.scene.control.ToggleButton;
 import javafx.scene.control.ToggleGroup;
+import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
 import javafx.stage.FileChooser;
+import javafx.util.Callback;
 public class MenuProfesorController {
 
 @SuppressWarnings("unused")
@@ -84,7 +95,6 @@ private MainProfesor main;
 	private ImageView imagenPregunta;
 	
 	private String radioButtonSelected = "1";
-	private final FileChooser fileChooser = new FileChooser();
 	private File imagenFile = null;
 	
 	@FXML
@@ -103,16 +113,22 @@ private MainProfesor main;
 		guardarUsuarioButton.onActionProperty().set(new EventHandler<ActionEvent>() {
 			public void handle(ActionEvent arg0) {
 				
-				Usuario usuario = new Usuario();
-				usuario.setNombre(nombreTextField.getText());
-				usuario.setDni(dniTextField.getText());
-				usuario.setPass(passwordTextField.getText());
-				usuario.setProfesor(profesorButton.isSelected());
+				if(!nombreTextField.getText().isEmpty() && !dniTextField.getText().isEmpty() && !passwordTextField.getText().isEmpty()){
 
-				if(ServiceLocator.getUsuarioServices().crearUsuario(usuario)){
+					Usuario usuario = new Usuario();
+					usuario.setNombre(nombreTextField.getText());
+					usuario.setDni(dniTextField.getText());
+					usuario.setPass(passwordTextField.getText());
+					usuario.setProfesor(profesorButton.isSelected());
+					
+					if(ServiceLocator.getUsuarioServices().crearUsuario(usuario)){
+						Utils.mensaje(AlertType.INFORMATION, "Correcto", "Confirmacion de Inserción", "Se ha registrado el nuevo usuario!");
+						limpiarFormularioAlumno();
+					}
+					else Utils.mensaje(AlertType.ERROR, "Error", "Error al crear", "Ha ocurrido un error al crear el usuario, comprueba los datos.!");
 				}
-				else{
-				}
+				else  Utils.mensaje(AlertType.ERROR, "Error", "Comprobación de los datos", "Faltan datos!");
+
 			}
 		});
 		
@@ -120,18 +136,55 @@ private MainProfesor main;
 		dniColumn.setCellValueFactory(cellData -> cellData.getValue().dniProperty());
 		nombreColumn.setCellValueFactory(cellData -> cellData.getValue().nombreProperty());
 		profesorColumn.setCellValueFactory(cellData -> cellData.getValue().profesorProperty());
-//		tablaUsuarios.getSelectionModel().selectedItemProperty().addListener(
-//	            (observable, oldValue, newValue) -> showUsuariosDetails(newValue));
+
+	    // cambiamos la propiedad de la columna e insertamos imagen según lo que reciba.             
+		profesorColumn.setCellFactory(new Callback<TableColumn<Usuario, Boolean>,TableCell<Usuario, Boolean>>(){        
+		    @Override
+		    public TableCell<Usuario, Boolean> call(TableColumn<Usuario, Boolean> param) {                
+		        TableCell<Usuario, Boolean> cell = new TableCell<Usuario, Boolean>(){
+		            @Override
+		            public void updateItem(Boolean item, boolean empty) {                        
+		                if(item!=null){ 
+		                	
+		                    HBox box= new HBox();
+		                    box.setAlignment(Pos.CENTER);
+
+		                    VBox vbox = new VBox();
+		                    vbox.setAlignment(Pos.CENTER);
+		                    
+		                    ImageView imageview = new ImageView();
+		                    imageview.setFitHeight(20);
+		                    imageview.setFitWidth(20);
+		                    
+		                    if(item)
+		                    	imageview.setImage(Images.CORRECT_IMAGE); 
+		                    else
+		                    	imageview.setImage(Images.INCORRECT_IMAGE);
+
+		                    box.getChildren().addAll(imageview, vbox); 
+		                    setGraphic(box);
+		                }
+		            }
+		        };              
+		        return cell;
+		    }
+		});
 
 		tablaUsuarios.setItems(ServiceLocator.getUsuarioServices().listarUsuarios());
 		
 		eliminarUsuarioButton.onActionProperty().set(new EventHandler<ActionEvent>() {
 			public void handle(ActionEvent event) {
-				Usuario usuario = tablaUsuarios.selectionModelProperty().get().getSelectedItem();
 				
-				if(ServiceLocator.getUsuarioServices().eliminarUsuario(usuario)){
-				}
-				else{
+				if(Utils.confirmacion("Confirmacion", 
+						"¿Realmente quiere eliminar este usuario?", 
+						"Si elimina este usuario no podrá recuperarlo y se eliminaran todos sus datos, incluido los resultados de sus test."))
+				{
+					Usuario usuario = tablaUsuarios.selectionModelProperty().get().getSelectedItem();
+					
+					if(ServiceLocator.getUsuarioServices().eliminarUsuario(usuario))
+						Utils.mensaje(AlertType.INFORMATION, "Correcto", "Se ha eliminado al usuario", "");
+					else  
+						Utils.mensaje(AlertType.ERROR, "Error", "Error al eliminar", "Ha ocurrido un error al eliminar el usuario, contacta con el administrador!");
 				}
 			}
 		});
@@ -143,8 +196,8 @@ private MainProfesor main;
 		imagenPregunta.onMouseClickedProperty().set(new EventHandler<Event>() {
 			public void handle(Event event) {
 				
-				//imagenFile = null;
-
+				imagenFile = null;
+				FileChooser fileChooser = new FileChooser();
                 fileChooser.getExtensionFilters().add(new FileChooser.ExtensionFilter("All Images", "*.*"));
                 fileChooser.getExtensionFilters().add(new FileChooser.ExtensionFilter("JPG", "*.jpg"));
                 fileChooser.getExtensionFilters().add(new FileChooser.ExtensionFilter("BMP", "*.bmp"));
@@ -178,23 +231,24 @@ private MainProfesor main;
 		guardarPreguntaButton.onActionProperty().set(new EventHandler<ActionEvent>() {
 			public void handle(ActionEvent arg0) {
 				
-				System.out.println(" el radio button indicado es : " + radioButtonSelected);
-				
-				Pregunta pregunta = new Pregunta();
-				pregunta.setEnunciado(enunciadoTextArea.getText());
-				pregunta.setPregunta1(pregunta1TextArea.getText());
-				pregunta.setPregunta2(pregunta2TextArea.getText());
-				pregunta.setPregunta3(pregunta3TextArea.getText());
-				pregunta.setRespuesta(radioButtonSelected);
-				pregunta.setImagen(imagenFile);
-				
-				if(ServiceLocator.getPreguntaServices().crearPregunta(pregunta)){
+				if(!enunciadoTextArea.getText().isEmpty() && !pregunta1TextArea.getText().isEmpty() && 
+					!pregunta2TextArea.getText().isEmpty() && !pregunta3TextArea.getText().isEmpty()){
+
+					Pregunta pregunta = new Pregunta();
+					pregunta.setEnunciado(enunciadoTextArea.getText());
+					pregunta.setPregunta1(pregunta1TextArea.getText());
+					pregunta.setPregunta2(pregunta2TextArea.getText());
+					pregunta.setPregunta3(pregunta3TextArea.getText());
+					pregunta.setRespuesta(radioButtonSelected);
+					pregunta.setImagen(imagenFile);
 					
-//					if(imagenFile != null)
-//						ServiceLocator.getPreguntaServices().guardarImagenDB(imagenFile);
+					if(ServiceLocator.getPreguntaServices().crearPregunta(pregunta)){
+						Utils.mensaje(AlertType.INFORMATION, "Correcto", "Se ha creado la pregunta", "La pregunta se ha creado correctamente!");
+						limpiarFormularioPregunta();
+					}
+					else  Utils.mensaje(AlertType.ERROR, "Error", "Error al crear", "Ha ocurrido un error al crear la pregunta, comprueba los datos!");
 				}
-				else{
-				}
+				else  Utils.mensaje(AlertType.ERROR, "Error", "Comprobación de los datos", "Faltan datos!");
 			}
 		});
 		
@@ -206,28 +260,55 @@ private MainProfesor main;
 		
 		eliminarPreguntaButton.onActionProperty().set(new EventHandler<ActionEvent>() {
 			public void handle(ActionEvent event) {
-				Pregunta pregunta = tablaPreguntas.selectionModelProperty().get().getSelectedItem();
 				
-				if(ServiceLocator.getPreguntaServices().eliminarPregunta(pregunta)){
-				}
-				else{
+				if(Utils.confirmacion("Confirmacion", 
+						"¿Realmente quiere eliminar esta pregunta?", 
+						"Si elimina esta pregunta no podrá recuperarla pero se mantendrán intacto los resultados de los test en la que se incluyó esta pregunta."))
+				{
+					Pregunta pregunta = tablaPreguntas.selectionModelProperty().get().getSelectedItem();
+					
+					if(ServiceLocator.getPreguntaServices().eliminarPregunta(pregunta))
+						Utils.mensaje(AlertType.INFORMATION, "Correcto", "Se ha eliminado la pregunta", "");
+					else  
+						Utils.mensaje(AlertType.ERROR, "Error", "Error al eliminar", "Ha ocurrido un error al eliminar la pregunta, contacta con el administrador!"); 
 				}
 			}
 		});
 		
 		tablaPreguntas.setRowFactory( tv -> {
-		    TableRow<Pregunta> row = new TableRow<>();
-			    row.setOnMouseClicked(event -> {
-			        if (event.getClickCount() == 2 && (! row.isEmpty()) ) {
-			        	
-			        	Pregunta rowData = row.getItem();
-	
-			        	//TODO aqui tengo ya los datos de las preguntas
-			        }
-			    });
+	    TableRow<Pregunta> row = new TableRow<>();
+		    row.setOnMouseClicked(event -> {
+		        if (event.getClickCount() == 2 && (! row.isEmpty()) ) {
+		        	
+		        	Pregunta rowData = row.getItem();
+		        	//TODO 
+		        }
+		    });
 		    return row ;
 		});
+	}	
+//		tablaPreguntas.getSelectionModel().selectedItemProperty().addListener(
+//				(observable, oldValue, newValue) -> showPreguntasDetails(newValue));
+//	}
+//	private void showPreguntasDetails(Pregunta newValue) {
+//		System.out.println(newValue.getPregunta1());
+//	}
+
+	private void limpiarFormularioAlumno(){
+		nombreTextField.setText("");
+		dniTextField.setText("");
+		passwordTextField.setText("");
 	}
+	private void limpiarFormularioPregunta(){
+		enunciadoTextArea.setText("");
+		pregunta1TextArea.setText("");
+		pregunta2TextArea.setText("");
+		pregunta3TextArea.setText("");
+		imagenPregunta.setImage(Images.INSERT_IMAGE);
+		imagenFile = null;
+	}
+	
+	
 
 	public void setMain(MainProfesor main) {
 		this.main = main;
